App/requirements.txt
streamlit
PyPDF2
langchain
langchain-google-genai
langchain-community
python-dotenv
langchain-huggingface
sentence-transformers
google-generativeai
faiss-cpu
pandas
joblib
shap
numpy
matplotlib
lime
xgboost
pytesseract
pdfplumber
PyMuPDF
Pillow
openai